Swapped position of function.
